Moved all components to new "Components" directory
